requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
gunicorn>=21.2
django-environ>=0.11
whitenoise>=6.6
dj-database-url>=1.3,<2.0  # Recommended for Django 5.x compatibility
Pillow>=10.0,<11.0  # For image fields in Django
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
